Added support for cover page in tile
Note that this is a first stab at adding a coverpage.

Changes are the coverpage (duh) and the t and h command line
variables to pass in info about the draft.

Note that I have yet to make this optional behaviour, and update
the man page.

But at least it outputs a cover page :)